Applied the patch supplied by Alec Thomas <alec@swapoff.org>.  This makes available multiple tagspaces which can be used to store different tags (for example, GalleryPlugin could use this as a tag provider instead of implementing it on its own).
Additionally, Alec simplified the name of the tags field in the table (tag instead of namespace).

This patch can handle an old database (it will migrate the data to the new table with a trac-admin upgrade).  All the macros have been converted to use the new tagging API.

More details about using the API is in api.py .


git-svn-id: https://trac-hacks.org/svn@1750 7322e99d-02ea-0310-aa39-e9a107903beb