GraphvizPlugin:
Corrected the self.log.debug statements that were printing out the req.args and req.base_url. They are not always available. Removed the reference to req.base_url since it was not being used and may not always be an attribute to req.  Fixes ticket:193. Added import for the inspect module. Fixes ticket:236.

git-svn-id: https://trac-hacks.org/svn@533 7322e99d-02ea-0310-aa39-e9a107903beb