WorkLogPlugin:
Reverts r2367. Refs #1710. Revert the time change code I put in as now this breaks it for me on my test system.
I do not understand why this is a problem as the exact same code is used in Trac itself (see attachment.py).
Please reopen if this remains an issue (note that it may require a clean install)

git-svn-id: https://trac-hacks.org/svn@2427 7322e99d-02ea-0310-aa39-e9a107903beb