closes #1913
After much waiting and thinking, I came to the conclusion that the best way to handle the dynamic statuses was to rebuild the reports each time they change.  To that end I have added code to check to see if the statuses have changed and then update the reports with the new statuses every time they change (requiring a trac-admin upgrade).  Rendering out the statuses and the linkifying javascript were changed to be compatible with this.

git-svn-id: https://trac-hacks.org/svn@2774 7322e99d-02ea-0310-aa39-e9a107903beb