TocMacro: Fixed a serious bug whereby the 0.11 version of the macro would keep a database connection open after execution.
This innocent-looking line...
{{{
self.formatter = formatter
}}}
... essentially attaches the transient `formatter` and its `formatter.db` connection to the instantiated plugin/component that lives for the lifetime of the environment (that in most setups gets cached for the duration of the process or until event like config change forces reload). Next time the macro executes it will free the old connection, and replace it with the new.

The problem continually ties up 1 connection of a process-wide default maximum number of 10. Many projects using the macro can get to the point where it locks the process due to no database connections being available - failing with `TimeoutError`. Particularly dangerous with embedded interpreter on Apache for Windows as it serves all requests using just a single process.

Code is changed to not persist the `formatter`, but instead pass it around as needed.

To all users of 0.11 version: Please upgrade!


git-svn-id: https://trac-hacks.org/svn@4366 7322e99d-02ea-0310-aa39-e9a107903beb