TracPastePlugin: Add finer-grained permissions, indexes and fix a bug.
PASTEBIN_USE is split into PASTEBIN_VIEW and PASTEBIN_CREATE. This
allows for finer-grained control of what a user is allowed to do.
An environment upgrade is required, which will also handle the
permission conversion; users which had PASTEBIN_USE will then have
PASTEBIN_VIEW and PASTEBIN_CREATE, and thus are allowed to do the
same things as before the upgrade.

In addition:
 * Add indexes to the {{{pastes}}} table, to improve speed of lookups.
 * The comment of our upgrade_environment implementation correctly
   states that we don't commit the changes, but we did it anyway.
   Fixed.


git-svn-id: https://trac-hacks.org/svn@4983 7322e99d-02ea-0310-aa39-e9a107903beb