TracPastePlugin: Allow to delete pastes.
Introduce new permission PASTEBIN_DELETE for that purpose. It's a
a separate permission that needs to be enabled explicitly for a
user, unless that user has PASTEBIN_ADMIN permission already.

A issue was introduced in r4983, where PASTEBIN_ADMIN inherited the
non-existant PASTEBIN_REPLY permission; that should be
PERMISSION_CREATE instead. This is now fixed.


git-svn-id: https://trac-hacks.org/svn@4984 7322e99d-02ea-0310-aa39-e9a107903beb