WikiTicketCalendarMacro: Add 1st try on i18n now (version 0.6.1).
This was attempted by adding a private translation module importing 
function from Trac's old translation utilities as well as from Babel 
directly.
Beware: This was bound to fail, since there is currently no known method 
to sanely attach translations to a single-file macro.
So this has to be reworked for the 0.12 branch later to get localized 
tool tips when moving over calendar day links.


git-svn-id: https://trac-hacks.org/svn@8051 7322e99d-02ea-0310-aa39-e9a107903beb