WikiTicketCalendarMacro: Remove tables from ticket description preview.
Even if supported by all tested browsers (looks really nice indeed),
this is not meant to work according to W3C (X)HTML specs. So better mask
it as done for images and links before than need to fix it later.
And some of the recently added CSS style definitions become obsolete too.

git-svn-id: https://trac-hacks.org/svn@8216 7322e99d-02ea-0310-aa39-e9a107903beb