DynamicFieldsPlugin: Merge changeset [9667] to `0.12` branch, refs #7729.
This removes the unsane "invalid_value" in favor of a clearer "hide_always"
option which also hides field on custom query page now.
Fixes small errors in `post_process_request` methode of the
!DynamicFieldsModule component too and stricter follows PEP8,
especially when writing logic 'and'/'or' before a line-break.

_Beware:_ This version is non-functional yet and need more changes,
so please stick to previous versions, if you're not developing yourself.

git-svn-id: https://trac-hacks.org/svn@9690 7322e99d-02ea-0310-aa39-e9a107903beb