Fix collapse widget after collapsing and reexpanding.
Really nothing to do with this ticket but I want a review of this minor
change.

This is a somewhat more complete fix than
http://code.google.com/p/jsgantt/issues/detail?id=95.

I don't know why the unicode character works on first draw but not when
redrawn.  This fixes both characters, the other fix doesn't change the
initial one.


git-svn-id: https://trac-hacks.org/svn@10029 7322e99d-02ea-0310-aa39-e9a107903beb