CustomFieldAdminPlugin: Finally got the "optional options" issue of select fields figured out. Closes #1834.
Turned out that creating and updating the field worked fine, but when changing field order all fields are deleted and recreated with new order - which did not properly account for the the possibility of the field being supplied with an `'optional'` key by the `TicketSystem`.

Wired up a test suite with tests that confirm that things work - and it should make it easier to look into issues in the future (in theory...).


git-svn-id: https://trac-hacks.org/svn@10368 7322e99d-02ea-0310-aa39-e9a107903beb