Add 'goal=' option to Gantt.
Similar to root=, if you specify a "goal", the chart shows all the
tickets required for the specified goal.  Follows the configurd 'pred'
field.

`goal=self` in a ticket description or comment displays the current
tickets predecessors.

`root`, `goal`, and `id` all add to the list of tickets to display but
do *not* interact.  If you specify `root=1,goal=2`, you get the
descendants of 1 and predecessors of 2 but not the descendants of the
predecessors or 2.

Author: Matt Sable

git-svn-id: https://trac-hacks.org/svn@10476 7322e99d-02ea-0310-aa39-e9a107903beb