AccountManagerPlugin: Restore full db deletion for deleted accounts, refs #9051.
Due to SQL injections protection - hey, it works ;-) - the SQL statements
responsible for purging user account information from the Trac db did not
work as expected, since changeset [10283].
I meant to have fixed this a while ago.  Obviously not, sorry.

I consider this a major bug, so will release another maintenance revision
for the stable release as well. Thanks for reporting this issue.

git-svn-id: https://trac-hacks.org/svn@10526 7322e99d-02ea-0310-aa39-e9a107903beb