TagsPlugin: Rework `TicketTagProvider`, refs #9059.
With i18n there's no point in hard-coding labels for tagged resource display.
It's not sensible to reinvent the wheel too, so `describe_tagged_resource`
method for tickets just calls the corresponding module in Trac core,
and we'll get good localization/translations of labels for free.

This is the final part of bundled suggestions contributed by Itamar Ostricher.
Thank you for all the invaluable ground work and inspiration.

git-svn-id: https://trac-hacks.org/svn@10682 7322e99d-02ea-0310-aa39-e9a107903beb