Fix conversion of estimate to work hours.  Refs #9042.
When scheduling was converted from days to hours, one conversion wasn't
addressed.

There are two factors that apply:

 * Time per estimate (how many scheduling units are in each unit of
estimate (and work)
 * Hours per day (how much work gets done each day)

The former affects the size of the task, the latter affects the
duration.  Both conversions were muddled up in `_workHours()` because we
used to convert estimate to days and wrongly concluced we then needed to
scale by hours per day to get work.

A failed case was an hour as a unit of estimate and a 6-hour work day
(or a day when only 6 hours of work get done).  With an estimate of 8,
days per estimate of .125, and hours per day of 6, this gave:

   hours = (8 * .125) * 6

Which converted the task for 8 hours to 6!

We first need to convert estimate to the unit of scheduling (hours) so
days_per_estimate becomes hours_per_estimate.  Computing work hours is
then just (estimate * hours per estimate).

When scheduling, we divide work hours by hours per day to get days.

git-svn-id: https://trac-hacks.org/svn@10809 7322e99d-02ea-0310-aa39-e9a107903beb