AccountManagerPlugin: Reimplement `PageCount` from WikiStatsPlugin, refs #9852.
Because this is a rather special macro I decided to include the functionality
in `Stats` wiki macro by using keyed parameters.   Added bonus is an optional
`prefix` key to only include pages with matching names.

If it seens preferable to retain the original macro as a short-hand, this could
still be done by adding it later on and treat it as an alias internally.

git-svn-id: https://trac-hacks.org/svn@11350 7322e99d-02ea-0310-aa39-e9a107903beb