TracAnnouncer: Adapt the template to data dict change in [12338], refs #10620.
Obviously I failed to track all my changes in that changeset, and especially
the move-and-modernize approach has fired back here -  sorry, and thanks to
Ryan J Ollos for spotting this as well as the right place for the fix.

I took the chance to improve template format like I'd do for others later on.
That's bound to break msgid's for translation, but I'll need to do a message
extraction for other recent i18n-relevant changes anyway.

git-svn-id: https://trac-hacks.org/svn@12356 7322e99d-02ea-0310-aa39-e9a107903beb