Dates derived from dependencies are not fixed.  Refs #9845.
If A has a fixed start and a successor, B, with a fixed end, an ALAP
schedule used to stretch A to run from its start to B's start.  Now it
shows A starting on schedule, ending after the expected duration, and
with slack before B's start.  (And vice versa for ASAP.)

git-svn-id: https://trac-hacks.org/svn@12427 7322e99d-02ea-0310-aa39-e9a107903beb