EXPERIMENTAL Add start, finish options for scheduler.  Refs #9845.
When doing an ASAP schedule of A and B where A is required for B but B
has a fixed start or finish date in the past, the Gantt chart would show
a backward-pointing dependency arrow.  This was because the schedule
started today by default (on the theory that new work can't start in the
past) so A started now and proceeded into the future but then B's
explict date made it show up before A.

Now you can say
`[[TracJSGanttChart(schedule=asap,start=2012-01-01,...)]]` to override
the default start of today.  (Similarly, you can use `finish` with
`alap` to override the default.

It may be better to let the default start be the beginning of time for
ASAP.  I'm not sure what the corresponding end of time for an ALAP
schedule would be.  And those may only be valuable if at least one task
in the project has a fixed date.

git-svn-id: https://trac-hacks.org/svn@12440 7322e99d-02ea-0310-aa39-e9a107903beb