AccountManagerPlugin: Add more HTML input fields and JS goodies, refs #8930 and Genshi:ticket:541.
Most of the existing configuration admin panel is visible in wizard pages
now, and I'm really looking forward to turn the wizard into a replacement for
that panel.

I decided to terminate final lines of JS code with a semicolon, because
a JS validator complained about it, and I remember earlier issues with IE
exactly for not having that final semicolon there.

See http://genshi.edgewall.org/ticket/541 for another problem, that I hit
during template development towards these changes.



git-svn-id: https://trac-hacks.org/svn@12587 7322e99d-02ea-0310-aa39-e9a107903beb