AccountManagerPlugin: Introduce code for changing session IDs consistently, refs #10900, #10901 and #10902.
There are actually three predecessors to this code (in creation time order):
RenameUsersScript, RenameTracUsersScript and RenameUserPlugin.  While the
scripts yield some hints on how to approach the task, the last one is the
most complete in terms of a real API, but can't be used for having the viral
GPL license attached, that would interfere and overpower the modified BSD,
that we use with this plugin.

So this is neither fresh code nor copied verbatim but something in-between.
I'm grateful and therefor grant every author of the aforementioned hacks my
respect for going ahead and sharing their solution with us.

You're welcome to join and help improve my attempt, that aims at obsoleting
all others by providing the best in Trac integration, flexibility,
reliability and performance.  Especially user ID changes in the Cc ticket
field and corresponding change records are not supported yet.

git-svn-id: https://trac-hacks.org/svn@12700 7322e99d-02ea-0310-aa39-e9a107903beb