AccountManagerPlugin: Move pager to user list in accounts admin panel, refs #809, #9946, #10682, #10745, #10754 and #10873.
This should fix broken display of extra-long user lists, and it makes
selections from user list useful for the clean-up page as requested.

Dumped access to anonymous session attributes for the current solution,
so these are no longer available for clean-up in the admin web-UI, but
I'll re-implement this later on, if it will be missed too much.

git-svn-id: https://trac-hacks.org/svn@12738 7322e99d-02ea-0310-aa39-e9a107903beb