Refine background ticket rescheduler.  Refs #9648.
Move the exclusion of milestone pseudo-tickets closer to where it makes
a difference (when we save the data).

This should be a no-op most of the time but without this the ticket set
sometimes has dangling references which makes the plugin fragile when
adding new function invocations in certain places.

git-svn-id: https://trac-hacks.org/svn@12763 7322e99d-02ea-0310-aa39-e9a107903beb